# ARROW-2121: [Python] Handle object arrays directly in pandas serializer.