--- a/python/benchmarks/convert_pandas.py
+++ b/python/benchmarks/convert_pandas.py
@@ -48,3 +48,23 @@ class PandasConversionsFromArrow(PandasConversionsBase):
 
     def time_to_series(self, n, dtype):
         self.arrow_data.to_pandas()
+
+
+class ZeroCopyPandasRead(object):
+
+    def setup(self):
+        # Transpose to make column-major
+        values = np.random.randn(10, 100000)
+
+        df = pd.DataFrame(values.T)
+        ctx = pa.default_serialization_context()
+
+        self.serialized = ctx.serialize(df)
+        self.as_buffer = self.serialized.to_buffer()
+        self.as_components = self.serialized.to_components()
+
+    def time_deserialize_from_buffer(self):
+        pa.deserialize(self.as_buffer)
+
+    def time_deserialize_from_components(self):
+        pa.deserialize_components(self.as_components)