The goal here is to get the best of both the `pandas_serialization_context` (speed at serializing pandas dataframes containing strings and other objects) and the `default_serialization_context` (correctly serializing a large class of numpy object arrays).

This PR sort of messes up the function `pa.pandas_compat.dataframe_to_serialized_dict`. Is that function just a helper function for implementing the custom pandas serializers? Or is it intended to be used in other places.

TODO in this PR (assuming you think this approach is reasonable):

- [x] remove `pandas_serialization_context`
- [x] make sure this code path is tested
- [x] double check performance/behavior

cc @wesm @pcmoritz @devin-petersohn