@wesm 2018-02-09
Well, we need to preserve the zero-copy pandas reads. Now that our ASV benchmarking setup has been rehabilitated we should be able to do that in this patch to verify performance