@robertnishihara 2018-02-09
Some performance numbers. The numbers are somewhat variable if you run the benchmarks multiple times.

```python
import pyarrow as pa
import pandas as pd
df = pd.DataFrame(data={str(i): [i, str(i)] for i in range(10 ** 6)})
```

Before this PR

```python
context = pa.pandas_serialization_context()

%time s = pa.serialize(df, context=context).to_buffer()  # 570ms
%time d = pa.deserialize(s, context=context)  # 485ms

%timeit s = pa.serialize(df, context=context).to_buffer()  # 482ms
%timeit d = pa.deserialize(s, context=context)  # 376ms
```

After this PR

```python
%time s = pa.serialize(df).to_buffer()  # 577ms
%time d = pa.deserialize(s)  # 672ms

%timeit s = pa.serialize(df).to_buffer()  # 467ms
%timeit d = pa.deserialize(s)  # 349ms
```