@robertnishihara 2018-02-09
Let's not merge this just yet, I'd like to brainstorm other approaches a little.