@robertnishihara 2018-02-09
Ok, I'm pretty happy with this now. @wesm @pcmoritz let me know if you have any comments.