@wesm 2018-02-19
Sorry for the delay, looking now, and may as well add a benchmark for zero-copy DataFrame while I'm at it