@wesm 2018-02-20
@robertnishihara would you mind enabling appveyor on your fork when you have a chance?