@robertnishihara 2018-02-20
Thanks @wesm I *think* I've enabled it now.