@wesm 2018-02-20
Merging this, since the last Appveyor build had passed